+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bug fixes
 
 -   Align YAML parsing with core Kubernetes supported YAML subset. (https://github.com/pulumi/pulumi-kubernetes/pull/690).
+-   Handle string values in the equalNumbers function. (https://github.com/pulumi/pulumi-kubernetes/pull/691).
 -   Properly detect readiness for Deployment scaled to 0. (https://github.com/pulumi/pulumi-kubernetes/pull/688).
 
 ## 0.25.5 (August 2, 2019)